+++ NEWS.md
@@ -11,8 +11,8 @@
 - `spatial_distribution_report` creates an HTML report showing the 
   location and nearest-neighbor relations of cells in a single field.
 - `read_components` reads component image files.
-- `count_within` counts the number of `from` cells having a `to` cell
-   within a given radius.
+- `count_within` and `count_within_batch` count the number of `from` 
+   cells having a `to` cell within a given radius.
 - `list_cell_seg_files` lists all cell seg data files in a folder.
 - Added vignettes
     - "Reading and Displaying inForm Image Files"